+++ src/constants/localStorageKeys.ts
@@ -1,0 +1,4 @@
+export const LOCAL_STORAGE_KEY = {
+  USER_INFO: "userInfo" as const,
+  AUTH_TOKEN: "authToken" as const,
+};
